move types for one file
