+++ src/components/favorites/favorites.tsx
@@ -5,8 +5,7 @@ import {Link} from 'react-router-dom';
 import FavoritesPlaceCard from '../favorites-place-card/favorites-place-card';
 import OffersOperation from '../../store/operations/offers-operation/offers-operation';
 import {getCitiesWithFavoriteOffers} from '../../store/reducers/favorite-offers-reducer/selectors';
-import {ICityWithFavoriteOffers} from '../../types/city-with-favorite-offers';
-import {IStore} from '../../types/store';
+import {ICityWithFavoriteOffers, IStore} from '../../types/types';
 
 interface FavoritesProps {
   citiesWithFavoriteOffers: ICityWithFavoriteOffers[],
