--- src/types/city-with-favorite-offers.d.ts
@@ -1,7 +1,0 @@
-import {IPlace} from './place';
-import {ICity} from './city';
-
-export interface ICityWithFavoriteOffers {
-  city: ICity,
-  offers: IPlace[]
-}
