--- src/types/city.d.ts
@@ -1,8 +1,0 @@
-export interface ICity {
-  location: {
-    latitude: number,
-    longitude: number,
-    zoom: number
-  },
-  name: string
-}
