--- src/types/offer.d.ts
@@ -1,29 +1,0 @@
-import {ICity} from './city';
-
-export interface IOffer {
-  bedrooms: number,
-  city: ICity,
-  description: string,
-  goods: string[],
-  host: {
-    avatar_url: string,
-    id: number,
-    is_pro: boolean,
-    name: string
-  },
-  id: number,
-  images: string[],
-  is_favorite: boolean,
-  is_premium: boolean;
-  location: {
-    latitude: number,
-    longitude: number,
-    zoom: number
-  },
-  max_adults: number,
-  preview_image: string,
-  price: number,
-  rating: number,
-  title: string,
-  type: `apartment` | `room` | `house` | `hotel`
-}
