--- src/types/review.ts
@@ -1,12 +1,0 @@
-export interface IReview {
-  comment: string,
-  date: string,
-  id: number,
-  rating: number,
-  user: {
-    avatar_url: string,
-    id: number,
-    is_pro: boolean,
-    name: string
-  }
-}
