--- src/types/user.d.ts
@@ -1,7 +1,0 @@
-export interface IUser {
-  id: number,
-  email: string,
-  name: string,
-  avatar_url: string,
-  is_pro: boolean
-}
